Rename Skill macro (Adv=Never) with Reliable Talent to "Ability" Skill macro (Adv=Never) with Reliable Talent
This does NOT use "selected" in the attribute calls, it MUST be in the Character's Ability macros. 

This is a variation of the Skill macro that Token Action Maker creates. Advantage is set up as Never and it includes automated support for the Rogue ability Reliable Talent. 

Also included is the current skill bonus in the dropdown menu. 

Other versions of this Skill macro will contain the other variations of Advantage